解决横向滚动和竖向滚动事件冲突

+++ app/src/main/java/com/example/commoncoordinatelayout/Util.java
@@ -32,7 +32,7 @@ public class Util {
     /**
      * 根据手机的分辨率从 dp 的单位 转成为 px(像素)
      */
-    public static int dip2px(Context context, float dpValue) {
+    public static int dip2px(Context context, int dpValue) {
         final float scale = context.getResources().getDisplayMetrics().density;
         return (int) (dpValue * scale + 0.5f);
     }
